requirements.txt
langchain
langgraph
pinecone
openai
google-generativeai
tqdm
requests
python-dotenv
langdetect
spacy
beautifulsoup4
fastapi
uvicorn
pydantic 